added flexbox to Home
