+++ src/App.js
@@ -2,7 +2,7 @@ import "./App.css";
 import Navbar from "./components/Navbar/Navbar.js";
 import { BrowserRouter as Router } from "react-router-dom";
 
-import Home from "./pages/Home.js";
+import Home from "./pages/Home/Home.js";
 
 function App() {
   return (
