+++ src/components/SidePanel/SidePanelElements.js
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const Sidebar = styled.section`
+  background: #fff;
+  border-style: solid;
+  border-color: #000;
+`;
+
+export const Intro = styled.section`
+  color: #525252;
+  align-items: left;
+`;
+
+export const Blurb = styled.section`
+  color: #525252;
+  align-items: left;
+`;
+
+export const Footer = styled.section`
+  color: #525252;
+  align-items: left;
+`;
