+++ src/pages/Home/HomeElements.js
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Main = styled.div`
+  color: #fff;
+  display: flex;
+`;
